backend/requirements.txt
Django==5.1.7
django-environ==0.12.0
psycopg2==2.9.10
djangorestframework==3.15.2
django_cors_headers==4.7.0
langchain_community==0.3.20
langchain-community==0.3.20
langchain==0.3.22
chromadb==0.6.3
langchain_chroma==0.2.2
langchain_huggingface==0.1.2
pypdf==5.4.0
pydantic==2.11.1
python-dotenv==1.1.0
langchain-openai==0.3.11